# Speedup `simplify_cfg` by avoiding Vec alloc in hot loop