@vaivaswatha 2022-10-22
That we're going through the entire program for replacing every value itself is algorithmically bad. Since we don't have def-use chains as part of the IR, a simpler solution is to keep all the necessary rewrites in a map and apply them at once. 

As part of my work for #2770, I'm going to be introducing changes to do this. So let's hope that'll help further. But in itself, this particular change looks fine to me.